requirements.txt
black==21.9b0
Django==3.2.7
djangorestframework==3.12.4
flake8==3.9.2
mypy==0.910
mypy-extensions==0.4.3
typing-extensions==3.10.0.2
